Fix build break because of type mismatching

+++ src/heap/gc-idle-time-handler.cc
@@ -47,19 +47,19 @@ void GCIdleTimeAction::Print() {
 void GCIdleTimeHandler::HeapState::Print() {
   PrintF("contexts_disposed=%d ", contexts_disposed);
   PrintF("contexts_disposal_rate=%f ", contexts_disposal_rate);
-  PrintF("size_of_objects=%" V8_PTR_PREFIX "d ", size_of_objects);
+  PrintF("size_of_objects=%" V8_PTR_PREFIX "d ", static_cast<int>(size_of_objects));
   PrintF("incremental_marking_stopped=%d ", incremental_marking_stopped);
   PrintF("can_start_incremental_marking=%d ", can_start_incremental_marking);
   PrintF("sweeping_in_progress=%d ", sweeping_in_progress);
   PrintF("mark_compact_speed=%" V8_PTR_PREFIX "d ",
-         mark_compact_speed_in_bytes_per_ms);
+         static_cast<int>(mark_compact_speed_in_bytes_per_ms));
   PrintF("incremental_marking_speed=%" V8_PTR_PREFIX "d ",
-         incremental_marking_speed_in_bytes_per_ms);
-  PrintF("scavenge_speed=%" V8_PTR_PREFIX "d ", scavenge_speed_in_bytes_per_ms);
-  PrintF("new_space_size=%" V8_PTR_PREFIX "d ", used_new_space_size);
-  PrintF("new_space_capacity=%" V8_PTR_PREFIX "d ", new_space_capacity);
+         static_cast<int>(incremental_marking_speed_in_bytes_per_ms));
+  PrintF("scavenge_speed=%" V8_PTR_PREFIX "d ", static_cast<int>(scavenge_speed_in_bytes_per_ms));
+  PrintF("new_space_size=%" V8_PTR_PREFIX "d ", static_cast<int>(used_new_space_size));
+  PrintF("new_space_capacity=%" V8_PTR_PREFIX "d ", static_cast<int>(new_space_capacity));
   PrintF("new_space_allocation_throughput=%" V8_PTR_PREFIX "d",
-         new_space_allocation_throughput_in_bytes_per_ms);
+         static_cast<int>(new_space_allocation_throughput_in_bytes_per_ms));
 }
